refactor(lab-l9yv.8): move SshSession to dispatch/deploy/ssh_session — keep pure SDK clean
SshSession, SshOptions, StrictHostKeyChecking, shell_quote, and SshError
all depend on tokio::process::Command and belong in the lab binary, not
in the lab-apis SDK. Move them to crates/lab/src/dispatch/deploy/ssh_session.rs.

crates/lab-apis/src/core/ssh.rs now contains only SshHostTarget and
parse_ssh_config — pure data types shared with extract. Zero references
to tokio::process::Command remain in lab-apis.

runner.rs import updated from lab_apis::core::ssh to super::ssh_session.
